Removed MovementRestrictions for US
These scenarios has movement restrictions for the US, that were supposed to be lifted when the US went to war. The lifting of restrictions was not firing and as the US normally goes to war on turn 1, I removed the restrictions. There were some minor restrictions on Russia & Japan that I also removed.